Create 2021EOS_Q8.c
You will generate the function to perform the operation shown starting with the code below
that includes the W1 and W2 weight arrays and that takes an input array. The input is onedimensional, size 6. W1 and W2 are two-dimensional: W1 is 6 by 7. W2 is 7 by 4. The output is
one-dimensional and of size 4.
Note: To multiply a Nx1 array A, by an MxN array B, to get an Mx1 array C:
C[0] = A[0] * B[0,0] + A[1] * B[0,1] + A[2] * B[0,2] + ... + A[N-1] * B[0,N-1]
C[1] = A[0] * B[1,0] + A[1] * B[1,1] + A[2] * B[1,2] + ... + A[N-1] * B[1,N-1]
...
C[M-1] = A[0] * B[M-1, 0] + A[1] * B[M-1, 1] + A[2] * B[M-1, 2]+ ... + A[N-1] * B[M-1, N-1]
7
Here is the function to finish:
#include <stdio.h>
// this function uses a one-dimensional input array of size 6, and
// puts values into the one-dimensional output array of size 4
void performAI(double *inArray, double *outArray) {
// these are the weighting arrays as in the diagram
double w1[6][7] = {
{0.795279571, 0.565454091, 0.569392801, 0.649519912, 0.311228459, 0.869033219, 0.963890145},
{0.261182548, 0.967901447, 0.015463096, 0.101966965, 0.454071297, 0.396147575, 0.853833996},
{0.976180547, 0.762522649, 0.223067359, 0.120228416, 0.710471648, 0.220771538, 0.052876278},
{0.173285965, 0.795507616, 0.258332188, 0.813302777, 0.528470338, 0.885245811, 0.190564347},
{0.14018923, 0.324797853, 0.012649753, 0.928397252, 0.048519668, 0.321836138, 0.360198988},
{0.063248883, 0.72395506, 0.606492812, 0.435057638, 0.462896967, 0.12061378, 0.28806367}};
double w2[7][4] = {
{0.036340161, 0.702081192, 0.406643568, 0.383400727},
{0.786459022, 0.627286192, 0.190417846, 0.259622675},
{0.996272492, 0.115783107, 0.922042574, 0.805576672},
{0.254649714, 0.818737484, 0.23760355, 0.884876231},
{0.587934606, 0.566762923, 0.254228386, 0.735145224},
{0.709219708, 0.815306359, 0.395073347, 0.191438772},
{0.743663242, 0.969784133, 0.055612809, 0.992284824}};
// your code here
};
Here is an example use of the function. Do NOT include the main() function in your submitted file:
double inData[] = {10, 11, 14, 51, 22, 24};
double outData[4];
performAI(inData, outData);