Create 2021EOS_Q9.c
Imagine we have a sorted array of int-type integers, a, in the following form:
a = {a1, a2, . . . , aN−1, aN }
where a1 ≤ a2 ≤ . . . ≤ aN−1 ≤ aN . We now define the following array from the sorted array a:
ad = {a1, aN , a2, aN−1, . . .}
In this array ad, the smallest item is followed by the greatest item, and then the second smallest
item and the second greatest item, and so on. We call this special array a Decussate-sorted array.
You are asked to write a function int *merge(int *size), which receives a series of Decussatesorted arrays of different lengths from user input, and returns a single-dimensional array that
merges them all together to one Decussate-sorted array. The merge() function has one parameter
size, which is used to return the size of the merged array to the calling function. Your merge
function is responsible for getting the array information from the user in the manner shown below.
Note: Your implementation should have no memory leaks. In other words, any dynamically allocated memory that you use inside the function should be freed. The returned array from the
merge() function will be freed by the calling function.
Hint: You may use a sorting algorithm if that simplifies your solution.
Here is an example main() function that can be used to test your work: